Boss scripting/testing
VERY basic placeholder animations for a boss fight, can be replaced with an actual snake, and animations when ready. scripting is mostly done, it requires, health does not function on the boss/nor do colliders yet OR its health bar, but the code is in place, and when the health does reach 0, a death  animation will play.