STORK UPDATE
Stork walks towards the player,
Does the 3 different attacks depend on distance, and if you leave the distance area it does the 3rd one after a time.
idles
If you are in the right position on the playforms, on an attack the bird falls down.  (Still buggy but visable) just change collider locations will probably be the answer
No health working yet (require stork to fall before i can do that.
Transitions between areas Works.
Need a way to limit travel.