Stork finished
Stork is finished, will require some animation work/sprite updates and of course a few tweaks later on, still but all the basics are done for it. Damage should not be taken while jumping over too. currently heads the only thing that does damage. though once we get a knockback thing in places legs can be easily change for that too.