Import Art assets, Work on Tutorial level(final)
I have created another scene in test build, Imported all currently complete assets for the scene, and have scattered them around our NEW tutorial level, also changed the prefab of save point to include a light when activated, and the prefab shoot to have some light with the fire as well. Turned Boar into a prefab as well so that any change done in the Boar test scene will reflect this scene too.