Bug Fixes and added
Fixed the freeze Bug at the end of the level had to research a weird bug that was freezing it. Layered the Boar correctly, added in its markers. Slightly modification to the typewriter prefab, changed scale to 0 as it was set to 100 Tutorial level 0.5 should be all good now. Speed up the animations on the main Menu.