Many updates
Intro comicbook style created with a word typing (modifyed from Davids script)
Includes a scipt aninmation button, so each page can be jumped past.
(also modifed the button so that it works with any animation... just have to type it in) can't find a way to access current animation apprently you cant? ) not worried about it. it works.
Fade in and out for each piece is done too, Replacement of images is straight froward. Timeing and words may need to change.

Scroll implmeneted, may require resizing depending on what we want.